feat: separate HomeFeed from HomePage
